renamed 'generate_benchmark' to 'driver'
